change type (striteri, strmapi)

+++ ft_striteri.c
@@ -6,7 +6,7 @@
 /*   By: susumuyagi <susumuyagi@student.42.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/23 13:22:15 by susumuyagi        #+#    #+#             */
-/*   Updated: 2023/05/23 13:28:21 by susumuyagi       ###   ########.fr       */
+/*   Updated: 2023/05/24 16:35:41 by susumuyagi       ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	ft_striteri(char *s, void (*f)(unsigned int, char *))
 {
-	size_t	i;
+	unsigned int	i;
 
 	if (!s || !f)
 		return ;
